cleanup(component): move LinkConverter to own file
